added new items in db, share inventory data to player

+++ gameserver/src/main/java/com/shnok/javaserver/model/item/ItemInfo.java
@@ -9,7 +9,7 @@ public class ItemInfo {
     private int objectId;
     private DBItem item;
     private int enchant;
-    private long count;
+    private int count;
     private int price;
     private byte equipped;
     private byte category;
